+++ app/javascript/lca/components/generic/RatingField.jsx
@@ -89,14 +89,14 @@ class RatingField extends Component<Props, State> {
       max,
       // eslint-disable-next-line no-unused-vars
       dontFocus,
-      ...fProps
+      ...otherProps
     } = this.props
     const { handleChange, handleFocus, handleBlur } = this
     const { value } = this.state
 
     return (
       <TextField
-        {...fProps}
+        {...otherProps}
         className={narrow ? classes.narrow : classes.field}
         type="number"
         name={trait}