+++ app/serializers/weapon_serializer.rb
@@ -2,7 +2,7 @@
 
 # app/serializers/weapon_serializer.rb
 class WeaponSerializer < CharacterTraitSerializer
-  attributes :name, :weight, :tags, :is_artifact,
+  attributes :name, :weight, :tags, :is_artifact, :notes,
              :ability, :attr, :damage_attr,
              :bonus_accuracy, :bonus_damage, :bonus_defense, :bonus_overwhelming
 end